+++ examples/testbed/net/prototype.rs
@@ -54,7 +54,7 @@ fn send_message_update_system(
         return;
     }
 
-    if let Some(_) = &mut changed_query.iter().into_iter().next() {
+    if changed_query.iter().into_iter().next().is_some() {
         broadcast_all_messages(net, all_messages_query);
     }
 }
@@ -73,11 +73,12 @@ fn send_cube_position_system(
         let pos = tx.0;
 
         let msg = TestbedMessage::CubePosition(pos.x(), pos.y(), pos.z());
-
-        net.broadcast(
-            &msg.encode()[..],
-            NetworkDelivery::UnreliableSequenced(Some(1)),
-        );
+        let _ = net
+            .broadcast(
+                &msg.encode()[..],
+                NetworkDelivery::UnreliableSequenced(Some(1)),
+            )
+            .unwrap();
     }
 }
 
@@ -86,17 +87,15 @@ fn send_create_message_system(
     net: Res<NetworkResource>,
     mut network_create_messages: ResMut<CreateMessages>,
 ) {
-    match &(*ci) {
-        ConnectionInfo::Client { server, .. } => {
-            for msg in &network_create_messages.messages {
-                net.send(
-                    *server,
-                    &TestbedMessage::CreateMessage(msg.clone()).encode()[..],
-                    NetworkDelivery::ReliableOrdered(Some(1)),
-                );
-            }
+    if let ConnectionInfo::Client { server, .. } = &(*ci) {
+        for msg in &network_create_messages.messages {
+            net.send(
+                *server,
+                &TestbedMessage::CreateMessage(msg.clone()).encode()[..],
+                NetworkDelivery::ReliableOrdered(Some(1)),
+            )
+            .expect("Create message failed to send");
         }
-        _ => {}
     }
 
     network_create_messages.messages.clear();
@@ -112,6 +111,7 @@ struct Players {
     names: std::collections::HashMap<Connection, String>,
 }
 
+#[allow(clippy::too_many_arguments)]
 fn handle_network_events(
     ci: Res<ConnectionInfo>,
     net: Res<NetworkResource>,
@@ -170,7 +170,7 @@ fn handle_introduction_event(
     players: &mut ResMut<Players>,
     client_update_events: &mut ResMut<Events<ClientUpdateEvent>>,
 ) {
-    net.send(
+    let _ = net.send(
         conn.addr,
         &TestbedMessage::Pong.encode()[..],
         NetworkDelivery::ReliableSequenced(Some(2)),
@@ -214,7 +214,7 @@ fn handle_sync_messages_event(
     sync_messages_events: &mut ResMut<Events<SyncMessagesEvent>>,
 ) {
     if ci.is_client() {
-        sync_messages_events.send(SyncMessagesEvent { messages: messages })
+        sync_messages_events.send(SyncMessagesEvent { messages })
     }
 }
 
@@ -224,16 +224,16 @@ fn broadcast_all_messages(net: Res<NetworkResource>, mut messages_query: Query<&
         messages.push(msg.clone());
     }
 
-    let sync_messages = TestbedMessage::SyncMessages { messages: messages };
+    let sync_messages = TestbedMessage::SyncMessages { messages };
 
-    net.broadcast(
+    let _ = net.broadcast(
         &sync_messages.encode()[..],
         NetworkDelivery::ReliableSequenced(Some(1)),
     );
 }
 
 fn start_server(addr: SocketAddr, mut net: ResMut<NetworkResource>) {
-    net.bind(addr).unwrap();
+    net.bind(addr).expect("We failed to bind to the socket.");
 }
 
 fn start_client(
@@ -242,13 +242,14 @@ fn start_client(
     server_addr: SocketAddr,
     mut net: ResMut<NetworkResource>,
 ) {
-    net.bind(addr).unwrap();
+    net.bind(addr).expect("We failed to bind to the socket.");
 
     net.send(
         server_addr,
         &TestbedMessage::Introduction(name.to_string()).encode()[..],
         NetworkDelivery::ReliableSequenced(Some(1)),
-    );
+    )
+    .expect("We failed to send our introduction message");
 }
 
 impl TestbedMessage {
@@ -256,18 +257,18 @@ impl TestbedMessage {
         if SERIALIZE_JSON {
             let encoded_json = serde_json::to_string(&self).unwrap();
             let bytes: Vec<u8> = encoded_json.as_bytes().to_vec();
-            return bytes;
+            bytes
         } else {
-            return bincode::serialize(&self).unwrap();
+            bincode::serialize(&self).unwrap()
         }
     }
 
     pub fn decode(bytes: &[u8]) -> TestbedMessage {
         if SERIALIZE_JSON {
             let encoded_json = std::str::from_utf8(bytes).unwrap();
-            return serde_json::from_str(&encoded_json).unwrap();
+            serde_json::from_str(&encoded_json).unwrap()
         } else {
-            return bincode::deserialize(bytes).unwrap();
+            bincode::deserialize(bytes).unwrap()
         }
     }
 }